requirements.txt
Flask
python-dotenv
redis
nanoid
